\name{bahamas}
\alias{bahamas}
\docType{data}
\title{bahamas map coordinates}
\description{a dataframe based on a .bna file downloaded from NOAA}
\format{A dataframe.
\describe{a dataframe (ncol=3) with longitude and latitude coordinates}
}
\details{A dataframe with longitude and latitude coordinates for
the Commonwealth of Bahamas}
\keyword{datasets}
\examples{str(bahamas)}
