\name{bahamas.sites}
\alias{bahamas.sites}
\docType{data}
\title{bahamas sites coordinates}
\description{a dataframe with names and coordinates of sampling sites (fictional) in Bahamas}
\format{A dataframe.
\describe{
\item{spec}{a dataframe (ncol=3) with names, longitude coordinates, and latitude coordinates}
 }
}
\details{A dataframe (ncol=3) with names, longitude coordinates, and latitude coordinates for fictional sites located in Bahamas}
\references{none so far}
\keyword{datasets}
\examples{str(bahamas.sites)}
