\name{bahamas.towns}
\alias{bahamas.towns}
\docType{data}
\title{bahamas town coordinates}
\description{a dataframe with names and coordinates of major towns in Bahamas}
\format{A dataframe.
\describe{
\item{spec}{a dataframe (ncol=3) with names, longitude coordinates, and latitude coordinates}
 }
}
\details{A dataframe (ncol=3) with names, longitude coordinates, and latitude coordinates for some of the major towns and cities of Bahamas}
\references{none so far}
\keyword{datasets}
\examples{str(bahamas.towns)}
