% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_palettes.R
\docType{data}
\name{map_palettes}
\alias{map_palettes}
\title{Complete list of pre-built color palettes}
\format{
An object of class \code{list} of length 1.
}
\usage{
map_palettes
}
\description{
Complete list of pre-built color palettes
}
\keyword{datasets}
