% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapratio.R
\name{mapratio}
\alias{mapratio}
\title{mapratio}
\usage{
mapratio(coords, map.width = 3, map.height = NULL)
}
\arguments{
\item{coords}{a data.frame with 3 columns based on a
GSHHS-NOAA .bna file}

\item{map.width}{a numerical value defining
map width in inches (default = 3)}

\item{map.height}{a numerical value defining map height.
If map height is provided, map width value is ignored and
the map dimensions are calculated based on height value.}
}
\value{
a vector with two values representing map dimensions (width, height)
}
\description{
Computes map dimensions for drawing maps based on GSHHS-NOAA Data
}
\details{
Computes map aspect ratio based on the x-axis (longitudinal) and y-axis
(latitudinal) ranges of the map, while also correcting for
longitude/latitude aspect ratio using mid-point latitude of the map.
Designed for GSHHS-NOAA .bna files (Atlas Boundary Files) with
shoreline coordinates defined as hierarchically arranged closed polygons.
Files accessible at \url{https://gnome.orr.noaa.gov/goods/tools/GSHHS/coast_subset}

Data need to be uploaded into an R data.frame to create an object
acceptable by the "mapratio" and "mapmaker" function.
Use this: object <- read.delim('filename.bna', sep=',', header=TRUE)

The resolution of a map drawn by mapmaker will depend on
the resolution chosen when downloading the .bna file from NOAA website.
For maps crossing the dateline longitude west values are > 180.
The map is always produced in the northward orientation.
}
\examples{
mapratio(bahamas)
mapratio(bahamas, map.height = 4)

}
