% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metro_map.R
\name{metro_map}
\alias{metro_map}
\title{Create a location based street map with the addition of metro lines}
\usage{
metro_map(
  location,
  color = map_palettes("Nightlight"),
  size = c(river = 1.2, railway = 0.5, highways = 0.7, main_streets = 0.5, small_streets
    = 0.1, metro = 1),
  linetype = c(river = "solid", railway = "dotdash", highways = "solid", main_streets =
    "solid", small_streets = "solid", metro = "solid"),
  transparency = c(river = 0.7, railway = 0.6, highways = 0.8, main_streets = 0.6,
    small_streets = 0.4, metro = 1),
  text_color = c(title = "white", subtitle = "white"),
  bounding_adj = 0.002,
  text_size = c(title = 20, subtitle = 10)
)
}
\arguments{
\item{location}{The name of the place you want to map via OSM query}

\item{color}{Named list for line colors}

\item{size}{Named list for line sizes}

\item{linetype}{Named list for line linetypes}

\item{bounding_adj}{Bounding Box Adjustment}

\item{text_size}{Text sizing for final plot output}
}
\description{
Create a location based street map with the addition of metro lines
}
