% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_map.R
\name{star_map}
\alias{star_map}
\title{Create a location based map of the stars}
\usage{
star_map(location, date, location_caption, color)
}
\arguments{
\item{location}{The name of the place you want to map via OSM query}

\item{date}{String format date of the time you would like the starmap}

\item{location_caption}{String placed for the location on the output plot}

\item{color}{Named list of colors for the star map}
}
\description{
Create a location based map of the stars
}
