% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/street_map.R
\name{street_map}
\alias{street_map}
\title{Create a location based street map}
\usage{
street_map(
  location,
  color = map_palette("Daylight"),
  size = c(river = 1.2, railway = 0.5, highway = 0.7, main_street = 0.5, small_street =
    0.1),
  linetype = c(river = "solid", railway = "dotdash", highway = "solid", main_street =
    "solid", small_street = "solid"),
  transparency = c(river = 1, railway = 1, highway = 1, main_street = 1, small_street =
    1),
  text_color = c(title = "white", subtitle = "white"),
  bounding_adj = 0.002,
  text_size = c(title = 20, subtitle = 10)
)
}
\arguments{
\item{location}{The name of the place you want to map via OSM query}

\item{color}{Named list for line colors}

\item{size}{Named list for line sizes}

\item{linetype}{Named list for line linetypes}

\item{bounding_adj}{Bounding Box Adjustment}

\item{text_size}{Text sizing for final plot output}
}
\description{
Create a location based street map
}
