% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapmaker-package.R, R/mapmaker.R
\docType{package}
\name{mapmaker}
\alias{mapmaker-package}
\alias{mapmaker}
\title{mapmaker: Creates a map based on NOAA coastline data}
\usage{
mapmaker(
  coords,
  pdf = FALSE,
  filename = "mymap.pdf",
  map.height = 5,
  scale.bar = TRUE,
  scale.width = NULL,
  scale.lwd = 1.5,
  scale.lat = NULL,
  scale.long = NULL,
  arr.north = TRUE,
  arr.lwd = 1.5,
  arr.cex = 0.8,
  arr.coord = NULL,
  sea.col = "lightskyblue1",
  coast.col = "khaki3",
  land.col = "khaki1",
  grayscale = FALSE,
  rscript = NULL
)
}
\arguments{
\item{coords}{a data.frame or matrix (ncol = 3) with geographic
coordinates based on a NOAA .bna file. The file}

\item{pdf}{logical (default = FALSE), determines if map is
plotted locally or saved as pdf file}

\item{filename}{character string (default = 'mymap.pdf')
defining name of the pdf file (ignored when pdf = FALSE)}

\item{map.height}{numerical value (default = 5) defining
the height of the map in inches (ignored if pdf = FALSE)}

\item{scale.bar}{logical (default = TRUE), determines if
the scale bar is plotted}

\item{scale.width}{numerical value (default = NULL),
the length of the scale bar (km)}

\item{scale.lwd}{numerical (default = 1.5), thickness of the
scale bar line}

\item{scale.lat}{numerical (default = NULL), latitude of the
scale bar (ignored if scale.long is not provided)}

\item{scale.long}{numerical (default = NULL), west longitude
of the scale bar (ignored if scale.long is not provided)}

\item{arr.north}{logical (default = TRUE), indicates
if N arrow should be plotted}

\item{arr.lwd}{numerical (default = 1.5), thickness of N arrow}

\item{arr.cex}{numerical (default = 0.8), the font size
of 'N' symbol of N arrow}

\item{arr.coord}{numerical vector (lenght=3) (default = NULL)
defining N arrow location (long, start lat, end lat)}

\item{sea.col}{color (default = 'lightskyblue1') for sea}

\item{coast.col}{color (default = 'khaki3') for coastline}

\item{land.col}{color name (default = 'khaki1') for land}

\item{grayscale}{logical (default = FALSE) overrides
all color definitions to plot gray-scale map}

\item{rscript}{R script (default = NULL) with additional
plot statements (e.g., mtext(), points(), etc.)}
}
\value{
a plot (map)
}
\description{
Package "mapmaker" creates PDF-formatted maps using NOAA data.

Returns a map with coordinates sourced from NOAA ".bna" data file
}
\details{
Generates a map based on .bna file downloaded from:
https://gnome.orr.noaa.gov/goods/tools/GSHHS/coast_subset.
A NOAA file can be uploaded using: read.delim('filename.bna', sep=',', header=TRUE)
If argument pdf = TRUE, a PDF file is generated with aspect ratios
based on mid-point latitude of the map.
A map will have a resolution given by the .bna file downloaded
by the user. For maps crossing the dateline longitude west
values are > 180. The map is always produced in the
northward orientation. Map aspect ratio is scaled only
for pdf outputs (pdf = TRUE) and local R plots (pdf = FALSE)
should be used for preview only. Map width < map height and
the difference in dimensions (aspect ratio)
increases with latitude.
}
\examples{

data(bahamas)
str(bahamas)


mapmaker(bahamas)


mapmaker(bahamas)

}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mjkowalewski/mapmaker}
  \item Report bugs at \url{https://github.com/mjkowalewski/mapmaker}
}

}
\author{
Michal Kowalewski \email{kowalewski@ufl.edu}
}
