% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapmaker.R
\name{mapmaker}
\alias{mapmaker}
\title{mapmaker}
\usage{
mapmaker(
  coords,
  pdf = FALSE,
  filename = "mymap.pdf",
  map.height = 5,
  scale.bar = TRUE,
  scale.width = NULL,
  scale.lwd = 1.5,
  scale.lat = NULL,
  scale.long = NULL,
  arr.north = TRUE,
  arr.lwd = 1.5,
  arr.cex = 0.8,
  arr.coord = NULL,
  sea.col = "lightskyblue1",
  coast.col = "khaki3",
  land.col = "khaki1",
  grayscale = FALSE,
  rscript = NULL
)
}
\arguments{
\item{coords}{a data.frame or matrix (ncol = 3) based on a
GSHHS-NOAA .bna file}

\item{pdf}{logical (default = FALSE), if map is saved to an
external pdf file or plotted locally}

\item{filename}{character string (default = 'mymap.pdf')
defining the name of the pdf output file (ignored when pdf = FALSE)}

\item{map.height}{numerical value (default = 5) defining
the height of the map in inches (ignored if pdf = FALSE)}

\item{scale.bar}{logical (default = TRUE), determines if
the scale bar is plotted}

\item{scale.width}{numerical value (default = NULL),
the length of the scale bar (km)}

\item{scale.lwd}{numerical (default = 1.5), thickness of the
scale bar line}

\item{scale.lat}{numerical (default = NULL), latitude of the
scale bar (ignored if scale.long is not provided)}

\item{scale.long}{numerical (default = NULL), west longitude
of the scale bar (ignored if scale.long is not provided)}

\item{arr.north}{logical (default = TRUE), indicates
if N arrow should be plotted}

\item{arr.lwd}{numerical (default = 1.5), thickness of N arrow}

\item{arr.cex}{numerical (default = 0.8), the font size
of 'N' symbol of N arrow}

\item{arr.coord}{numerical vector (lenght=3) (default = NULL)
defining N arrow location (long, start lat, end lat)}

\item{sea.col}{the color used for sea areas (default = 'lightskyblue1')}

\item{coast.col}{the color used for coastline (default = 'khaki3')}

\item{land.col}{the color used for land areas (default = 'khaki1')}

\item{grayscale}{logical (default = FALSE), if the grayscale map
should be drawn (overrides other color definitions)}

\item{rscript}{R script (default = NULL) with additional
plot statements (e.g., mtext(), points(), etc.)}
}
\value{
either a plot (pdf = FALSE) or a pdf file (pdf = TRUE)
}
\description{
Draw Maps using GSHHS-NOAA Data
}
\details{
Generates a map using GSHHS-NOAA data that can be downloaded at
\url{https://gnome.orr.noaa.gov/goods/tools/GSHHS/coast_subset}.
The data download as .bna files (Atlas Boundary Files) with
shoreline coordinates defined as hierarchically arranged closed polygons.


Data need to be uploaded into an R data.frame to create an object
acceptable by the mapmaker function.


object <- read.delim('filename.bna', sep=',', header=TRUE)


The resolution of a map drawn by mapmaker will depend on
the resolution chosen when downloading the .bna file from NOAA website.
For maps crossing the dateline longitude west values are > 180.
The map is always produced in the northward orientation.
The function computes map aspect ratio based on the mid-point
latitude of a given map. The aspect ratio only applies to pdf-formatted
figures produced by mapmaker function when argument pdf = TRUE.
The scaling is not applied When the map is previewed in R plot window
(pdf = FALSE).
}
\examples{
mapmaker(bahamas)

}
