% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapmaker.R
\name{mapmaker}
\alias{mapmaker}
\title{mapmaker}
\usage{
mapmaker(
  coords,
  pdf = FALSE,
  filename = "mymap.pdf",
  map.height = 5,
  custom.dim = NULL,
  scale.bar = TRUE,
  scale.width = NULL,
  scale.lwd = 1.5,
  scale.lat = NULL,
  scale.long = NULL,
  arr.north = TRUE,
  arr.lwd = 1.5,
  arr.cex = 0.8,
  arr.coord = NULL,
  sea.col = "lightskyblue1",
  coast.col = "khaki3",
  land.col = "khaki1",
  grayscale = FALSE,
  transp = 0.5,
  rscript = NULL
)
}
\arguments{
\item{coords}{a data.frame or matrix (ncol = 3) based on a
GSHHS-NOAA .bna file}

\item{pdf}{logical, determines if map is saved to an
external pdf file or plotted locally (default = FALSE)}

\item{filename}{character string defining the name of the pdf
file (default = 'mymap.pdf'). Ignored when pdf = FALSE)}

\item{map.height}{numerical value defining the height of the map
in inches (default = 5). Ignored when pdf = FALSE}

\item{custom.dim}{vector with two values defining map
dimensions in inches (width and height) (default = NULL).}

\item{scale.bar}{logical, determines if the scale bar is plotted
(default = TRUE). If scale.bar = FALSE, other scale parameters are ignored}

\item{scale.width}{numerical value, the length of the scale bar
in kilometers (default = NULL)}

\item{scale.lwd}{numerical value, the scale bar line
width (default = 1.5)}

\item{scale.lat}{numerical value, latitude of the scale bar
(default = NULL). Ignored if scale.long not provided}

\item{scale.long}{numerical value, west longitude endpoint
of the scale bar (default = NULL). Ignored if scale.lat not provided.}

\item{arr.north}{logical, indicates if N arrow is plotted (default = TRUE).
When arr.north = FALSE all arrow parameters are ignored.}

\item{arr.lwd}{numerical value, thickness of N arrow (default = 1.5)}

\item{arr.cex}{numerical value, the font size of 'N' symbol
(default = 0.8)}

\item{arr.coord}{numerical vector with three values defining
N arrow location (long, start lat, end lat) (default = NULL).
If not provided, N arrow is plotted in the upper right corner of the map.}

\item{sea.col}{the color used for sea areas (default = 'lightskyblue1')}

\item{coast.col}{the color used for coastline (default = 'khaki3')}

\item{land.col}{the color used for land areas (default = 'khaki1')}

\item{grayscale}{logical (default = FALSE), if the grayscale map
should be drawn (overrides other color definitions)}

\item{transp}{numerical value for land color transparency (default = 0.5)}

\item{rscript}{R script (default = NULL) with additional
plot statements (e.g., mtext(), points(), etc.)}
}
\value{
either a plot (pdf = FALSE) or a pdf file (pdf = TRUE)
}
\description{
Draw Maps using GSHHS-NOAA Data
}
\details{
Generates a map using GSHHS-NOAA data that can be downloaded at
\url{https://gnome.orr.noaa.gov/goods/tools/GSHHS/coast_subset}.
The data download as .bna files (Atlas Boundary Files) with
shoreline coordinates defined as hierarchically arranged closed polygons.


Data need to be uploaded into an R data.frame to create an object
acceptable by the mapmaker function.


object <- read.delim('filename.bna', sep=',', header=TRUE)


The resolution of a map drawn by mapmaker will depend on
the resolution chosen when downloading the .bna file from NOAA website.
For maps crossing the dateline longitude west values are > 180.
The map is always produced in the northward orientation.
The function computes map aspect ratio based on the mid-point
latitude of a given map. The aspect ratio only applies to pdf-formatted
figures produced by mapmaker function when argument pdf = TRUE.
The scaling is not applied When the map is previewed in R plot window
(pdf = FALSE).
}
\examples{
mapmaker(bahamas)
mapmaker(bahamas, grayscale = TRUE, rscript = {
points(bahamas.towns[, 2:3], pch = 24, cex = 0.6, col = 'black', bg = 'blue')
text(bahamas.towns[, 2:3], cex = 0.5, col = 'blue', bahamas.towns[,1], pos = 2, offset=0.15)
points(bahamas.sites[, 2:3], pch = 21, cex = 1.3, col = 'black', bg = 'white')
text(bahamas.sites[, 2:3], cex = 0.4, col = 'black', bahamas.sites[,1])
})

}
